Start of a command menu

+++ Övning 5 - Garage/Vehicles/Boat.cs
@@ -10,8 +10,8 @@ namespace Exercise_5_Garage.Vehicles
     public class Boat : Vehicle
     {
         public int Length { get; protected set; }
-        public int Draft { get; protected set; }
-        public Boat(string brandAndModel, string color, int numberOfWheels, PowerType powerSource, string registrationNumber, int length, int draft) : base(brandAndModel, color, numberOfWheels, powerSource, registrationNumber)
+        public double Draft { get; protected set; }
+        public Boat(string brandAndModel, string color, int numberOfWheels, PowerType powerSource, string registrationNumber, int length, double draft) : base(brandAndModel, color, numberOfWheels, powerSource, registrationNumber)
         {
             Length = length;
             Draft = draft;
